% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_sketch.R
\name{count_sketch}
\alias{count_sketch}
\title{CountSketch}
\usage{
count_sketch(data, m)
}
\arguments{
\item{data}{(n times d)-dimensional matrix of data}

\item{m}{subsample size that is less than n}
}
\value{
(m times d)-dimensional matrix of data
}
\description{
Provides a subsample of data using CountSketch
}
