% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sketch.R
\name{sketch}
\alias{sketch}
\title{Sketch}
\usage{
sketch(data, m, method = "unif")
}
\arguments{
\item{data}{(n times d)-dimensional matrix of data}

\item{m}{subsample size that is less than n}

\item{method}{method for sketching}
}
\value{
(m times d)-dimensional matrix of data
}
\description{
Provides a subsample of data using sketches
}
\examples{
fullsample <- matrix(rnorm(100), nrow = 20,ncol = 5)
subsample <-  sketch(fullsample, 10)
subsample <-  sketch(fullsample, 10, "unif__with_replacement")
subsample <-  sketch(fullsample, 10, "CountSketch")
subsample <-  sketch(fullsample, 10, "fft")

}
\references{
Sokbae Lee, S and Ng, S. (2020). An Econometric Perspective on Algorithmic Subsampling.
Annual Review of Economics, 12:1, 45-80.
}
